Fixing imports and renaming package to follow pep 8
