+++ yodatools/dataloader/view/WizardSummaryPageView.py
